Update LICENSE
I have to use GPL because I want to use a GPL'd library.  BOO on the FSF!  I want the FREEDOM of choosing which open source license I want to use!  But hey, if you want to use their code, you've got to follow their rules.  They're as bad about forcing terms on you as a proprietary license.  I prefer Apache or MIT.